Restructured project to include ssml.cast utility
