+++ pyproject.toml
@@ -4,7 +4,4 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "ssml"
-version = "0.0.1"
-
-[project.scripts]
-cast = "ssml:cast"
+version = "1.0.0"
